# doc/api.yaml
...
openapi: 3.0.0 
info:  
  title: WasaPhoto 
  description:  |  
    This API allows you to manage the 
    data available in the project 
  version: 0.0.1 
servers:
  - url: http://localhost:3000
paths:  
  /session: 
    post: 
      tags: ["doLogin"] 
      summary: Logs in the user 
      description: |- 
        If the user does not exist, it will be created, 
        and an identifier is returned. 
        If the user exists, the user identifier is returned. 
      operationId: doLogin 
      requestBody:
        $ref: "#/components/requestBodies/userDetailsNickname"

      responses:  
        '201': 
          description: User log-in action successful 
          content: 
            application/json: 
              schema: 
                  $ref: "#/components/schemas/userId"
              example:
                userId: 111112
        '411': { $ref: "#/components/responses/UsernameNotValid" }
        '500': {$ref: "#/components/responses/ErrorServerExecution"}  
  /users/{userId}/username: 

    put: 
          tags: ["setMyUserName"] 
          summary: Set Username of User selected by his id  
          description: | 
            If the user does not exist, error is returned. 
            If the user exists, his username is changed. 
          operationId: setMyUserName 
          security: 
            - BearerAuth: [] 
          parameters:  
            - $ref: "#/components/parameters/userId" 
          requestBody:  
            $ref: "#/components/requestBodies/userDetailsNickname"

          responses:  
            '200': 
              description: UserName changed
              content: 
                application/json: 
                  schema: 
                    $ref: "#/components/schemas/scUserName"
                  example:
                    username: "Marco12"

            '400': { $ref: "#/components/responses/UserIdNotFound" }
            '401': { $ref: "#/components/responses/UserNotAuthorized" }
            '411': { $ref: "#/components/responses/UsernameNotValid" }
            '500': {$ref: "#/components/responses/ErrorServerExecution"}  
  /users/{userId}/followUser/{userId2}: 
    parameters:  
      - $ref: "#/components/parameters/userId" 
      - $ref: "#/components/parameters/userId2"     

    put: 
          tags: ["followUser"] 
          summary: Follow another user 
          description: | 
            If the user identified by userId or userId2 does not exist, error is returned. 
            If the user identified by userId2 exists, user identified by userId follow user identified by userId2 if user2 hasn't banned him. 
          operationId: followUser 
          security: 
            - BearerAuth: [] 

          responses:  
            '200': 
              description: Now you follow the other user 
              content: 
                application/json: 
                  schema: 
                    $ref: '#/components/schemas/scUserName' 
                  example:
                    username: "Luigi11" 

            '400': { $ref: "#/components/responses/UserIdNotFound" }                  
            '403': {$ref: "#/components/responses/UserIdBanned"}
            '404': {$ref: "#/components/responses/UserId2NotFound"} 
            '405': {$ref: "#/components/responses/userId2Banned"}
            '500': {$ref: "#/components/responses/ErrorServerExecution"}  
            '502': {$ref: "#/components/responses/FailedAutentication"}  
            '503': {$ref: "#/components/responses/NotAllowed"}  

    delete:
          tags: ["unfollowUser"] 
          summary: Unfollow another user 
          description: | 
            If the user identified by userId or userId2 does not exist or user doesn't follow userId2, error is returned. 
            If user follow userId2, user identified by userId unfollow user identified by userId2. 
          operationId: unfollowUser 
          security: 
            - BearerAuth: [] 

          responses:  
            '200': 
              description: Now you don't follow the other user 
              content: 
                application/json: 
                  schema: 
                    $ref: '#/components/schemas/scUserName' 
                  example:
                    username: "Luigi11" 

            '400': { $ref: "#/components/responses/UserIdNotFound" }                  
            '403': {$ref: "#/components/responses/UserIdBanned"}   
            '402': {$ref: "#/components/responses/UnfollowedUser"}    
            '404': {$ref: "#/components/responses/UserId2NotFound"}
            '405': {$ref: "#/components/responses/userId2Banned"}   
            '500': {$ref: "#/components/responses/ErrorServerExecution"}  
            '502': {$ref: "#/components/responses/FailedAutentication"}  
...
  /user/{userId}/photo/{photoId}:  
      parameters:
        - $ref:  '#/components/parameters/photoId'
        - $ref: '#/components/parameters/userId'
      delete:
          tags: ["deletePhoto"] 
          summary:  Remove photo from his profile 
          description: | 
            If the user identified by userId or photoId  does not exist or user doesn't posted the photo, error is returned. 
            If user had posted the photo and the photoId exists, he removes his  post
          operationId: deletePhoto 
          security: 
            - BearerAuth: [] 

          responses:  
            '200': 
              description: Remove photo
              content: 
                application/json: 
                  schema: 
                    $ref: '#/components/schemas/errorSchema' 
                  example:
                    result: "Done" 

            '401': { $ref: "#/components/responses/UserNotAuthorized" }
            '407': {$ref: "#/components/responses/photoNotFound"}              
            '500': {$ref: "#/components/responses/ErrorServerExecution"}         


      get:
          tags: ["getPhoto"]
          summary:  Return single photo
          description: Return the single requested photo in binary format.
          operationId: getPhoto
          security:
            - BearerAuth: []
          responses:
            '200':
              description: The requested photo in binary format.
              content:
                application/json:
                  schema:
                    $ref: "#/components/schemas/imageD"
                  example:
                    image: "34qfni23"

            '400': { $ref: "#/components/responses/UserIdNotFound" }   
            '403': {$ref: "#/components/responses/UserIdBanned"}
            '404': {$ref: "#/components/responses/UserId2NotFound"} 
            '405': {$ref: "#/components/responses/userId2Banned"} 
            '407': {$ref: "#/components/responses/photoNotFound"}   
            '500': {$ref: "#/components/responses/ErrorServerExecution"}  
...
  /users:
    get:
      tags: ["getUserId"]
      summary: Get userId from his username
      description:  |
        If user identified by userId in authentication doesn't exist or user identified by the username passed doesn't exist, error is returned.
        If user identified by userId in authentication exist and user identified by the username passed exist, the userId of the user is ruterned.
      operationId: getUserId
      security:
        - BearerAuth: []
      parameters:
        - name: username
          in: query
          schema:
            $ref: "#/components/schemas/username"
          required: true
          description: Username of the user to search 
      responses:
          '201': 
            description: User log-in action successful 
            content: 
              application/json: 
                schema: 
                    $ref: "#/components/schemas/userId"
                example:
                  userId: 111112
          '404': { $ref: "#/components/responses/UsernameNotFound" }    
          '400': { $ref: "#/components/responses/UserIdNotFound" }    
          '500': {$ref: "#/components/responses/ErrorServerExecution"} 
          '411': { $ref: "#/components/responses/UsernameNotValid" }
...
components:  
  responses:
...
    UsernameNotValid:
     description: Username not valid
    FailedAutentication:
      description: Failed authentication
    UsernameNotFound:
      description: Searched username not found
    UserIdNotFound:
      description: User id not found
    UserNotAuthorized:
      description: User not authorized 
    UserIdBanned:
      description: User banned by User2
    userId2Banned:
      description: User banned User2
    UserId2NotFound:
      description: Userid2 not found    
    ErrorServerExecution: 
      description: Error server execution     
    UnfollowedUser:
      description: You don't follow userId2 
...
    photoNotFound:
      description: Photo Id not found
...
    NotAllowed:
      description: "Operation not allowed"
  schemas: 
    username: 
      type: string 
      description: User's UserName 
      example: "Franchino_12"
      pattern: '^[a-zA-Z0-9_.-]*$' 
      minLength: 6 
      maxLength: 16  
    userId:
      type: object 
      description: User log-in action successful 
      properties: 
        identifier: 
          $ref: '#/components/schemas/identifier'
    identifier: 
      type: integer
      description: User Id 
      example: 1 
...
    image:
      description: The photo file
      type: string
      format: binary
      minLength: 1
      maxLength: 10485760 
      pattern: "((.|\n)*)" 
    imageD:
      type: object
      description: "Return codified photo"
      properties:
        image:
          $ref: "#/components/schemas/image"
...
    errorSchema: 
      description: Error example
      type: object 
      properties: 
        Message: 
          $ref: "#/components/schemas/errorDescription" 
    errorDescription: 
      type: string  
      description: Status
      example: Success 
      pattern: '^[a-zA-Z0-9]+$' 
      minLength: 4
      maxLength: 50 
...
    scUserName:
      type: object
      description:  "Describe userName" 
      properties:
        username:
          $ref: "#/components/schemas/username"  
...
  requestBodies: 
...
    userDetailsNickname:
          description: User details Nickname
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/scUserName"
  parameters:  
    userId: 
      in: path 
      name: userId 
      required: true 
      description: User Id 
      schema: 
        type: integer 
        description: "User Id" 
        example: 86 
    userId2: 
      in: path 
      name: userId2 
      required: true 
      description: Second userId
      schema: 
        type: integer 
        description: "Second userId" 
        example: 100 
    photoId: 
      in: path 
      name: photoId 
      required: true 
      description: Single id photo
      schema: 
        type: integer 
        description: "Single id photo"
        example: 15368 
...
  securitySchemes: 
    BearerAuth: 
      description: login only as your user id 
      type: http 
      scheme: bearer 